form page divide into separate component & user route-view for tabs
