+++ resources/js/pages/forms/show/analytics.vue
@@ -1,0 +1,34 @@
+<template>
+  <div>
+    <h3 class="font-semibold mt-4 text-xl">
+      Form Analytics (last 30 days)
+    </h3>
+    <form-stats :form="form"/>
+  </div>
+</template>
+
+<script>
+import FormStats from '../../../components/open/forms/components/FormStats'
+
+export default {
+  components: {FormStats},
+  props: {
+    form: { type: Object, required: true }
+  },
+
+  metaInfo() {
+    return {title: 'Form Analytics'}
+  },
+
+  data: () => ({
+  }),
+
+  mounted() {},
+
+  computed: {},
+
+  methods: {
+    
+  }
+}
+</script>
